+++ Source/Core/src/com/shtrih/jpos/fiscalprinter/receipt/NullReceipt.java
@@ -74,7 +74,7 @@ public class NullReceipt implements FiscalReceipt {
         notOpenedException();
     }
 
-    public void printRecItem(String description, long price, int quantity,
+    public void printRecItem(String description, long price, double quantity,
             int vatInfo, long unitPrice, String unitName) throws Exception {
         notOpenedException();
     }
@@ -112,7 +112,7 @@ public class NullReceipt implements FiscalReceipt {
         cancelled = true;
     }
 
-    public void printRecVoidItem(String description, long amount, int quantity,
+    public void printRecVoidItem(String description, long amount, double quantity,
             int adjustmentType, long adjustment, int vatInfo) throws Exception {
         notOpenedException();
     }
@@ -121,7 +121,7 @@ public class NullReceipt implements FiscalReceipt {
         notOpenedException();
     }
 
-    public void printRecItemFuel(String description, long price, int quantity,
+    public void printRecItemFuel(String description, long price, double quantity,
             int vatInfo, long unitPrice, String unitName, long specialTax,
             String specialTaxName) throws Exception {
         notOpenedException();
@@ -156,7 +156,7 @@ public class NullReceipt implements FiscalReceipt {
         notOpenedException();
     }
 
-    public void printRecItemVoid(String description, long price, int quantity,
+    public void printRecItemVoid(String description, long price, double quantity,
             int vatInfo, long unitPrice, String unitName) throws Exception {
         notOpenedException();
     }
@@ -167,13 +167,13 @@ public class NullReceipt implements FiscalReceipt {
     }
 
     public void printRecItemRefund(String description, long amount,
-            int quantity, int vatInfo, long unitAmount, String unitName)
+            double quantity, int vatInfo, long unitAmount, String unitName)
             throws Exception {
         notOpenedException();
     }
 
     public void printRecItemRefundVoid(String description, long amount,
-            int quantity, int vatInfo, long unitAmount, String unitName)
+            double quantity, int vatInfo, long unitAmount, String unitName)
             throws Exception {
         notOpenedException();
     }
